fix: 하단 장바구니옵션창 여닫힘에 대한 상태참조 위치 변경
resolved: #126

- 옵션창의 store상태참조 위치를 하위 컴포넌트인 BottomCartOption컴포넌트로 이동
